Heh.  The android clients behave poorly if the username is blank
Apparently they repeatedly slam the server (once every couple of seconds)
getting a 400 error back.  The client code should be changed to realize
this error is fatal and it should stop retrying.  I bet it is really
hammering device battery as well. (attn @arthurbenemann).  The server
side 'fix' is to include a 1 minute delay before returning this error
msg.  We'll need to fix the java client lib eventually.